Module data_streaming_service::data_notification

Structs

AccountsWithProofRequest

A request for fetching account states.

DataNotification

A single data notification with an ID and data payload.

EpochEndingLedgerInfosRequest

A client request for fetching epoch ending ledger infos.

NumberOfAccountsRequest

A client request for fetching the number of accounts at a version.

PendingClientResponse

A pending client response where data has been requested from the network and will be available in client_response when received.

TransactionOutputsWithProofRequest

A client request for fetching transaction outputs with proofs.

TransactionsWithProofRequest

A client request for fetching transactions with proofs.

Enums

DataClientRequest

A request that has been sent to the Diem data client.

DataPayload

A single payload (e.g. chunk) of data delivered to a data listener.

Type Definitions

NotificationId

A unique ID used to identify each notification.